Added sanity checks for wipe option, fixed execute option

+++ mimk.py
@@ -160,8 +160,8 @@ def run_command(command_str, undo=False):
 
 
 # Main program
-mimk_version = '1.25'
-mimk_date = '2019-03-24'
+mimk_version = '1.26'
+mimk_date = '2019-08-06'
 global args
 parser = argparse.ArgumentParser(description='mimk - Minimal make')
 parser.add_argument('target', help='Target configuration file')
@@ -257,7 +257,7 @@ if args.execute:
                 if 'TARGET' in target_attr:
                     if not any(target['TARGET'] == target_attr['TARGET'] for target in targets):
                         targets.append(target_attr)
-    args.execute.extend(new_execute_list)
+    args.execute = new_execute_list
 
 # Build dir paths
 build_dir = os.path.join('build', config['BUILD'])
@@ -265,7 +265,7 @@ hashes_file = '.hashes.json'
 hashes_path = os.path.join(build_dir, hashes_file)
 
 # Wipe build database
-if args.wipe:
+if args.wipe and os.path.isdir(build_path):
     try:
         os.remove(hashes_path)
         shutil.rmtree(build_dir)
@@ -316,7 +316,7 @@ for index, target in enumerate(targets):
     obj_dir = os.path.join(build_dir, config['OBJPATH'])
 
     # Wipe object folder
-    if args.wipe:
+    if args.wipe and os.path.isdir(obj_path):
         try:
             shutil.rmtree(obj_dir)
         except Exception:
